Move `CurveApproxSegment` to dedicated module
The new module hierarchy might seem a bit convoluted. It's designed to
mirror the module hierarchy I have in a local branch, where I've
developed all of this a bit further.
