+++ crates/fj-core/src/algorithms/approx/curve.rs
@@ -1,0 +1,5 @@
+//! Curve approximation
+
+mod segment;
+
+pub use self::segment::CurveApproxSegment;
