+++ crates/fj-core/src/algorithms/approx/mod.rs
@@ -1,5 +1,6 @@
 //! Approximation of objects
 
+pub mod curve;
 pub mod cycle;
 pub mod edge;
 pub mod face;
